Rebuild composer file based on cf-ex-drupal.
In an effort to closer align with the cloud.gov Drupal example, this adds a
handful of dependencies. It looks like the previous code was also built with a
generator, so this shouldn't be a big issue. All of the previous Drupal
modules should still be present.

Notably, however, this *does* declare that we're using PHP 7.1.