Add settings files.
These are largely auto-generated, but we don't want to recreate them every
time an instance restarts, so we'll commit them to the repository.

The exception is the settings.cf.php file, which configures the database
connection, flysystem, and the hash salt based on values injected by cloud.gov
services. The values can be missing (e.g. if running locally), and a local
settings file can be used instead.