Add a "cronish" worker to call cron every 15m.
Cloud.gov doesn't offer time-based task execution yet; luckily, it's very easy
to imitate with a loop in bash. This'll execute the `cron:execute` command
roughly every 15m (there's a slight skew due to amount of time processing that
command).

It also moves app configuration up a level in the manifest, as most of these
configs are shared.