Set logs to store indefinitely.
As part of our SSR, we're claiming to store logs for at least a year. We've
also noted that the logs are stored (for now) in the database.

As @fureigh pointed out, this is a frequent performance bottleneck for Drupal
sites. Unfortunately, there's not enough runway to create a longer term
solution (like writing to syslog) as we know from our users that they want to
see subsets of the logs within the application. We're kicking that can down
the road a bit.