Implement approach to module/theme removal.
This should solve a complicated problem. As we add modules to our site,
they're rolled out via configuration synchronization. This'll run the
installation of new modules, including setting up database tables.
Unfortunately, removing modules isn't as simple as deleting the PHP lib and
deactivating the plugin. Modules and themes need to be fully uninstalled,
which will remove their content from the database and perform other sorts of
cleanup. Unfortunately, to do that, we need to have the PHP lib around to run
the cleanup.

So! What this does is add a step to our bootstrap script to run those
uninstall scripts when we deploy code. We will *keep* the PHP lib and a
reference to its Drupal uninstallation around until we're confident all
environments have been updated. At that point, we can remove the lib from
composer and remove the entry in this bootstrap script.