Remove the Masquerade module
The Masquerade module (drupal.org/project/masquerade) provides the
ability for site administrators (or anyone with sufficient
permission) to switch to the perspective of a different user
without invoking a password. It has been useful for local
development but it 1) isn't relevant to the functionality we've
identified for the beta launch and 2) shouldn't be enabled on a
production site in any case.