Switch to multi-buildpack to install mysql.
As drush needs a mysql client, we need to use the apt-buildpack to install one
(the repository already had a configuration for apt). As this means we need
both apt and php, we'll use the multi-buildpack.

Note that this'll require tweaking our ATO language a bit, as we're now more
responsible for updating PHP.