Make footer menus' block titles display
By default, the USWDS theme doesn't render block titles for footer
menu blocks. This commit overrides that. Now, if a footer menu
block has a title, that title will display.

Note: the default heading level for block titles is h2. This
commit wraps the block titles in h3 tags instead, because 1) that
more closely approximates the appearance of the mockup and 2) more
importantly, there doesn't seem to be a semantic reason for using
h2 instead. Consistency is valuable, though, so I flag the
divergence here in case it's worth reconsidering.

https://github.com/18F/nsf/issues/190