Replace flysystem with s3fs.
s3fs will allow us to host our "public" files (e.g. compressed CSS/JS) on S3,
a feature flysystem lacks. We need this so that those files are saved between
deployments (and across instances).

This removes S3 from the docker-compose config so that the s3fs config is only
set when in cloud.gov (or mimicking locally).

Generated twig files are no longer stored in the "public" dir (as that'd touch
S3), so we ignore them in a gitignore.