Uninstall the Contact module
The Contact module is the core module that creates personal
contact forms as well as the sitewide contact form. Because we're
not currently using that functionality, we can turn it off by
uninstalling the Contact module, which includes deleting its
configuration files.

By handling #205 at the module level, we enable a future developer
to reenable the module and have it work as expected, while slightly
improving performance in the meantime (by not loading the Contact
module on every page load even though it's not really being used).

http://github.com/18F/nsf/issues/205