Refactor the 'homepage' to be a regular page plus blocks
Previously, the main page was generated by a view (as a 'page'
display). This commit adds a blank page node at the same URL,
deletes the no-longer-used page display of that view, and updates
the block view to 1) match the appearance of the earlier approach
and 2) be Tumblr-specific.

The Tumblr feeds are aggregator sources 1 and 2, hence the filter
for "less than or equal to 2".

https://github.com/18F/nsf/issues/155
https://github.com/18F/nsf/issues/162