Bump to Drupal 8.5.6.
Drupal core issued a security release today that bumps up their dependent
libraries: https://www.drupal.org/SA-CORE-2018-005

To satisfy, this updates Drupal and three of the symfony libraries it depends
on. In one of the patch releases sine 8.5.3, Drupal also modified two of the
files it auto-creates; we've included those changes here as well.